Refactored to make way for updated autoscaling algorithm (some of it was just general logical cleanup)
